Main

Description

Programa que implementa el juego Wordle. Contiene dos modos de juego:

Modo Mente Maestra: El jugador tiene que adivinar la palabra que ha sido seleccionada por el sistema.

Modo Descrifrador: La computadora debe adivinar la palabra que el jugador selecciona.

Documentation

contains

Arguments

:: Eq a

=> a

-> [a]
Caracter a buscar.

-> Bool
True si el string contiene el caracter especificado , False en caso contrario.

Funcion que verifica si un string contiene el caracter especificado.

splitAllWords

Arguments

:: [String]
Lista de palabras a separar.

-> Int
posicion de la palabra en el array de palabras.

-> Int
tamaño de la lista de palabras.

-> [[[Char]]]
Lista de palabras separadas.

-> [[[Char]]]
Retorna la lista de palabras separadas.

Funcion que separa todas las palabras de la lista grande en caracteres, para poder comparar caracter por caracter.

mayuscula

Arguments

:: String
String a transformar.

-> String
Retorna el string transformado todo en upperCase.

Funcion que transforma todas las letras de un string en mayusculas.

filterByToros

Arguments

:: [([Char], Int)]
Array de tuplas con las letras que son T y su posicion

-> Int
Posicion del Array de palabras

-> Int
Tamaño del Array de palabras

-> [[[Char]]]
Lista de palabras filtradas por contener T en el string de Evaluacion

-> [[[Char]]]
Retorna la lista de palabras filtradas por contener T en el string de Evaluacion

Funcion que filtra la lista de todas las palabras, por aquellas que posean las letras (Aquellas con T) que componen al String de evaluacion, en la misma posicion . Y retorna un array con las palabras que cumplen con la condicion.

filterByVacas

Arguments

:: [([Char], Int)]
Array de tuplas con las letras que son T y su posicion

-> Int
Posicion del Array de palabras

-> Int
Tamaño del Array de palabras

-> [[[Char]]]
Lista de palabras filtradas por contener V en el string de Evaluacion

-> [[[Char]]]
Retorna la lista de palabras filtradas por contener V en el string de Evaluacion

Funcion que filtra la lista de todas las palabras previamente filtradas T, poraquellas que posean las letras (Aquellas con V) que componen al String de evaluacion, en diferente posicion . Y retorna un array con las palabras que cumplen con la condicion.

asignarPuntaje

Arguments

:: [Char]
Letra a evaluar

-> Float
Retorna el valor de evaluacion de la letra

Funcion que retorna el valor de evaluacion segun la ponderacion especificada en el PDF a partir de una letra.

obtenerPuntajeW

Arguments

:: Int
Posicion del array del string

-> [[Char]]
String a evaluar caracter por caracter

-> ([[Char]], Float)
Tupla con el array de caracteres y su evaluacion

-> ([[Char]], Float)
Retorna la tupla con el string y su evaluacion

Funcion que retorna una tupla con un array de Chars y su respectiva evaluacion segun el puntaje.

obtenerPuntajeTot

Arguments

:: Int
Posicion del array del string

-> Int
Tamaño del array de palabras

-> [[[Char]]]
Array de palabras a evaluar

-> [([[Char]], Float)]
Array de tuplas con el string y su evaluacion

-> [([[Char]], Float)]
Retorna el array de tuplas con el string y su evaluacion

Funcion que retorna una lista de tuplas con arrays de Chars y su respectiva evaluacion segun el puntaje.

revisarToros

Arguments

:: Int
Posicion del array del string.

-> [[Char]]
String a evaluar caracter por caracter (la palabra adivinada por la computadora).

-> [[Char]]
String de evaluacion ingresada por el usuario.

-> [([Char], Int)]
Array de tuplas con el caracter correspondiente a una T y su posicion en la palabra adivinada por la computadora.

-> [([Char], Int)]
Retorna un array de tuplas con el caracter correspondiente a una T y su posicion en la palabra de evaluacion ingresada por el usuario.

Funcion que toma el string de evaluacion ingresado por el usuario, la palabra adivinada por la computadora y un array vacio de tuplas y retorna un array de tuplas con el caracter correspondiente a una T junto a la posicion de la misma.

revisarVacas

Arguments

:: Int
Posicion del array del string.

-> [[Char]]
String a evaluar caracter por caracter (la palabra adivinada por la computadora).

-> [[Char]]
String de evaluacion ingresada por el usuario.

-> [([Char], Int)]
Array de tuplas con el caracter correspondiente a una V y su posicion en la palabra adivinada por la computadora.

-> [([Char], Int)]
Retorna un array de tuplas con el caracter correspondiente a una V y su posicion en la palabra de evaluacion ingresada por el usuario.

Funcion que toma el string de evaluacion ingresado por el usuario, la palabra adivinada por la computadora y un array vacio de tuplas y retorna un array de tuplas con el caracter correspondiente a una V junto a la posicion de la misma.

revisionV

Arguments

:: Int
Posicion del array del string

-> [[Char]]
String a evaluar caracter por caracter (la palabra ingresada por el usuario)

-> [[Char]]
Array de T de la palabra ingresada por el usuario

-> [Char]
String de evaluacion de la palabra ingresada por el usuario

-> [Char]
String de evaluacion de la palabra ingresada por el usuario con las V cambiadas por -

Funcion que recibe la palabra ingresada por el usuarios la lista de letras que tiene T ancladas en el string de evaluacion y el string de evaluacion de la palabra ingresada por el usuario y verifica que no existan V si existe una T anclada a la letra en otro lado de la palabra.Si existe un caso como el anterior, entonces divide el string, cambia la V por - y vuelve a unirla, se llama recursivamente. -

revisar

Arguments

:: Int
Posicion del array del string.

-> [[Char]]
Palabra ingresada por el usuario al adivinar.

-> [[Char]]
Palabra random generada por la computadora, que es la palabra a adivinar por el usuario.

-> [Char]
String de evaluacion de la palabra ingresada por el usuario.

-> [[Char]]
Letras que el usuario acerto y por lo tanto seran T del string de evaluacion.

-> [Char]

Funcion que recibe la palabra ingresada por el usuario, la palabra random generada por la computadora, una lista vacia que sera la contenedora del string de evaluacion y otra lista contenedora que sera de la letras que tengan T ancladas a ellas en el string de evalacion.

initMenteMaestra

Arguments

:: Int
Turno actual del juego.

-> String
Palabra random escogida por la computadora.

-> IO ()

Funcion encarga de inicializar el modo mente maestra del juego de vacas y toros, se llama recursivamente hasta que el usuario ingrese una palabra valida para poder se evaluada y se detiene luego del turno 6, indicandole al usuario que ha perdidO y cual era la palabra random. En caso de que el usuario adivine correctamente se le indica que ha ganado.

initDecifrador

Arguments

:: Int
Turno actual del juego.

-> String
Palabra random escogida por la computadora.

-> [String]
Lista de todas las palabras.

-> Int
Tamaño de la lista de palabras.

-> IO ()

main :: IO ()